updated sqltables to use bigserial as primary key.

+++ app/src/sql/test/task.sql
@@ -14,7 +14,7 @@
 DROP TABLE task;
 CREATE TABLE task(
  ID     BIGSERIAL PRIMARY KEY,
- UserID               INTEGER,
+ UserID               BIGSERIAL,
  Memo                 VARCHAR(1000),
  RepeatIntervalInDays BIGINT,
  TaskLength           BIGINT,
